Komol parpixojayev

+++ src/infrastructure/base/base.service.ts
@@ -1,5 +1,8 @@
 import {
   DeepPartial,
+  Entity,
+  FindOptionsWhere,
+  FindOptionsWhereProperty,
   ObjectLiteral,
   Repository,
 } from 'typeorm';
@@ -17,21 +20,31 @@ export class BaseService<CreateDto, UpdateDto, Entity extends ObjectLiteral> {
     return this.baseRepo;
   }
   // ============================ CREATE ============================
+
   async create(dto: CreateDto): Promise<ISuccessRes> {
+    // make schema
     const entity = this.baseRepo.create(dto as DeepPartial<Entity>);
-    const data = this.baseRepo.save(entity);
-    return successRes(entity);
+
+    // save database
+    const data = await this.baseRepo.save(entity);
+
+    // unhhide is_deleted
+    delete data.is_deleted;
+
+    // return success
+    return successRes(data);
   }
 
   // ============================ FIND ALL ============================
+
   async findAll(options?: IFindOption<Entity>): Promise<ISuccessRes> {
     const data = await this.baseRepo.find({ ...options });
     return successRes(data);
   }
 
   // ============================ FIND BY ============================
-  async findOneBY(options?: IFindOption<Entity>): Promise<ISuccessRes> {
 
+  async findOneBY(options?: IFindOption<Entity>): Promise<ISuccessRes> {
     // find option
     const data = await this.baseRepo.find({
       select: options?.select || {},
@@ -49,13 +62,13 @@ export class BaseService<CreateDto, UpdateDto, Entity extends ObjectLiteral> {
   }
 
   // ============================ FIND BY ID ============================
+
   async findOneById(
     id: number,
     options?: IFindOption<Entity>,
   ): Promise<ISuccessRes> {
-
     // find by id
-    if(this.baseRepo.metadata.name=='admin'){
+    if (this.baseRepo.metadata.name == 'admin') {
       if (id == config.SUPERADMIN.ID) {
         throw new NotFoundException('You could not show Super Admin');
       }
@@ -79,9 +92,26 @@ export class BaseService<CreateDto, UpdateDto, Entity extends ObjectLiteral> {
     return successRes(data);
   }
 
-  // ============================ UPDATE ============================
-  async update(id: number, dto: UpdateDto): Promise<ISuccessRes> {
+  // ============================ FIND BY ID ON REPOSITORY ============================
 
+  async findByIdRepository<T extends ObjectLiteral>(
+    repository: Repository<T>,
+    id: number,
+  ): Promise<ISuccessRes> {
+    const data = await repository.findOne({
+      where: { id } as unknown as Entity,
+    });
+    if (!data) {
+      throw new NotFoundException(
+        `not found this id => ${id} on ${String(repository.metadata.name).split('Entity')[0]}`,
+      );
+    }
+    return successRes(data);
+  }
+
+  // ============================ UPDATE ============================
+
+  async update(id: number, dto: UpdateDto): Promise<ISuccessRes> {
     // check id
     await this.findOneById(id);
 
@@ -94,10 +124,18 @@ export class BaseService<CreateDto, UpdateDto, Entity extends ObjectLiteral> {
   }
 
   // ============================ DELETE ============================
-  async remove(id: number): Promise<ISuccessRes> {
 
+  async remove(id: number): Promise<ISuccessRes> {
     // check id
-    await this.findOneById(id);
+    const exist = await this.baseRepo.findOne({
+      where: { id } as unknown as Entity,
+    });
+
+    if (!exist) {
+      throw new NotFoundException(
+        `not found this id => ${id} on ${String(this.baseRepo.metadata.name).split('Entity')[0]}`,
+      );
+    }
 
     // delete
     await this.baseRepo.delete(id);
@@ -105,15 +143,17 @@ export class BaseService<CreateDto, UpdateDto, Entity extends ObjectLiteral> {
     // return success
     return successRes({});
   }
+
   // ============================ SOFT DELETE ============================
 
   async softDelete(id: number): Promise<ISuccessRes> {
-
     // check id
     await this.findOneById(id);
 
     // soft delete is is_deleted=true
-    await this.baseRepo.update(id, { is_deleted: true } as any);
+    await this.baseRepo.update(id, {
+      is_deleted: true,
+    } as unknown as QueryDeepPartialEntity<Entity>);
 
     return successRes({});
   }
